tools/services/ai-behavioral-nudge-engine/requirements.txt
# Web Framework and Server
fastapi==0.111.0
uvicorn==0.30.1

# Data handling and processing
pandas==2.2.2
numpy==1.26.4

# Machine Learning / AI libraries (adjust based on your actual needs)
scikit-learn==1.5.0
# tensorflow==2.x.x # Uncomment and specify version if using TensorFlow
# pytorch==2.x.x # Uncomment and specify version if using PyTorch

# Other utilities
python-dotenv==1.0.1 # For loading .env files if needed
# Add any other libraries your src/*.py files depend on

